Pulses driver splitted
