+++ radio/src/targets/horus/board_horus.h
@@ -165,6 +165,10 @@ void SDRAM_Init(void);
 #define IS_INTERNAL_MODULE_ON()   (GPIO_ReadInputDataBit(INTMODULE_PWR_GPIO, INTMODULE_PWR_GPIO_PIN) == Bit_SET)
 #define IS_EXTERNAL_MODULE_ON()   (GPIO_ReadInputDataBit(EXTMODULE_PWR_GPIO, EXTMODULE_PWR_GPIO_PIN) == Bit_SET)
 #define IS_UART_MODULE(port)      (port == INTERNAL_MODULE)
+
+// void setupPulses(uint8_t port);
+// void setupPulsesPPM(uint8_t port);
+// void setupPulsesPXX(uint8_t port);
 void init_no_pulses(uint32_t port);
 void disable_no_pulses(uint32_t port);
 void init_ppm( uint32_t module_index );
